FriendListSideBar: some css changes
